requirements.txt
cryptography==40.0.2
Django==4.2.1
Pillow==9.5.0
Pillow==9.5.0
pycryptodome==3.17
